requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.6
anthropic==0.26.1
resend==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
email-validator==2.1.1
